scripts/requirements.txt
backcall==0.2.0
decorator==4.4.2
flake8==3.8.3
gitdb==4.0.5
GitPython==3.1.7
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
six==1.15.0
smmap==3.0.4
toml==0.10.1
traitlets==4.3.3
wcwidth==0.2.5
